repo/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
Flask-Login==0.6.3
bcrypt==4.2.0
cryptography==42.0.8
pyotp==2.9.0
qrcode==7.4.2
python-dotenv==1.0.1
flask-wtf==1.2.1
